implement health monitor for edge detail

+++ IVCNetMaui/Models/Status/ProcessStatus.cs
@@ -8,12 +8,12 @@ public class ProcessStatus
     public int Pid { get; set; }
     public DateTime? LastExitTime { get; set; }
     public int? LastExitCode { get; set; } = -1;
-    public int CpuTotal { get; set; }
+    public double CpuTotal { get; set; }
     public double CpuUsed { get; set; }
     public long RamPeakWorking { get; set; }
     public long RamWorking { get; set; }
-    public int RamPaged { get; set; }
-    public int RamPeakPaged { get; set; }
+    public long RamPaged { get; set; }
+    public long RamPeakPaged { get; set; }
     public int Threads { get; set; }
     public int Handles { get; set; }
 }
